Fixes more issues reported by sonarqube

+++ swe-common-core/src/main/java/org/vast/swe/DataInputStreamLI.java
@@ -237,8 +237,8 @@ public class DataInputStreamLI extends FilterInputStream implements DataInputExt
     {
         int total = 0;
         int cur = 0;
-
-        while ((total < n) && ((cur = (int) in.skip(n - total)) > 0))
+        
+        while ((total < n) && ((cur = (int) in.skip((long)(n-total))) > 0))
         {
             total += cur;
         }
